+++ CHANGELOG.md
@@ -8,11 +8,13 @@
 - `flz.get_datasets` can return `Dataset` objects instead of path strings if 
   `return_paths=False`
 - New `OnixData` class to handle Onix data
+- `get_flexilims_session` can now re-use token from a previous session
 
 ### Minor
 - `CameraDataset` metadata can also be `.yml`, not only `.txt`.
 - `Dataset.format(mode='yaml')` ensure yaml compatibility. (path to str, tuple to list, 
   etc...)
+- `add_experimental_session` can be done with `parent_id` (or `parent_name`).
 
 ### Bugfixes
 - Fix [#68](https://github.com/znamlab/flexiznam/issues/68). Dataset.format returns